pyproject.toml
[project]
name = "as-tool-part2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
